requirements.txt
numpy
pickle
tqdm
faiss-cpu
